[refactor] (bitcoin) move common HD test to helper.ts
